+++ app/helpers/h.rb
@@ -1,0 +1,3 @@
+# Short alias
+# Example : H.number(123)
+H = Agilibox::AllHelpers